requirements.txt
beautifulsoup4==4.12.3
httpx==0.27.2
numpy==2.1.3
openai==1.54.4
python-dotenv==1.0.1
transformers==4.46.0